requirements/requirements-testing.txt
coverage
#coveralls
django-dynamic-fixture
django-nose
freezegun
mock
psycopg2
